# openapi.yaml
openapi: 3.0.0
info:
  title: User API
  description: API to manage user data.
  version: 1.0.0

paths:
  /users:
    post:
      summary: Create a new user.
      description: Creates a new user with the provided details.
      operationId: createUser
      requestBody:
        description: User details to create a new user.
        content:
          application/json:
            schema:
              type: object
              properties:
                uuid:
                  type: string
                  example: "123e4567-e89b-12d3-a456-426614174000"
                email:
                  type: string
                  format: email
                  example: user@gmail.com
                name:
                  type: string
                  example: Username
                urlProfilePhoto:
                  type: string
                  format: uri
                  example: https://firebasestorage.googleapis.com/v0/profile_picture_user.jpg
                provider:
                  type: string
                  example: google.com
              required:
                - email
                - name
                - provider
                - uuid
      responses:
        '201':
          description: User created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: Email or UUID already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ERROR_CREATE_USER"
              examples:
                EMAIL_ERROR:
                  value:
                    message: Email already exists
                    error: Conflict
                    statusCode: 409              
                UUID_ERROR:
                  value:
                    message: UUID already exists
                    error: Conflict
                    statusCode: 409  
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error"
                  error:
                    type: string
                    example: "Internal Server Error"
                  statusCode:
                    type: integer
                    example: 500
  /users/{user_uuid}:
    get:
      summary: Get user by UUID.
      description: Retrieves user information by UUID.
      operationId: getUserByUuid
      parameters:
        - in: path
          name: user_uuid
          required: true
          description: UUID of the user to retrieve.
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614173210"
      responses:
        '200':
          description: User found successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
                  error:
                    type: string
                    example: "Not Found"
                  statusCode:
                    type: integer
                    example: 404
        '505':
          description: Server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Server error"
                  error:
                    type: string
                    example: "Internal Server Error"
                  statusCode:
                    type: integer
                    example: 505

  /users/{user_uuid}/location:
    patch:
      summary: Update user location.
      description: Updates the latitude and longitude for a user based on their ID.
      operationId: updateUserLocation
      parameters:
        - in: path
          name: user_uuid
          required: true
          description: UUID of the user to update.
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        description: Location data to update the user's location.
        content:
          application/json:
            schema:
              type: object
              properties:
                latitude:
                  type: number
                  format: float
                  example: 34.6037
                longitude:
                  type: number
                  format: float
                  example: 58.3816
              required:
                - latitude
                - longitude
      responses:
        '200':
          description: User location updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
                  error:
                    type: string
                    example: "Not Found"
                  statusCode:
                    type: integer
                    example: 404
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error"
                  error:
                    type: string
                    example: "Internal Server Error"
                  statusCode:
                    type: integer
                    example: 500

  /users/{user_email}/failed-attempts:
    patch:
      summary: Increment the number of failed login attempts for a user.
      description: Increases the failed login attempt counter and locks the account if needed.
      operationId: increaseFailedAttempts
      parameters:
        - in: path
          name: user_email
          required: true
          description: Email of the user to update failed attempts.
          schema:
            type: string
            example: "user@gmail.com"
      responses:
        '200':
          description: Failed attempts updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
        '403':
          description: Account is locked due to multiple failed attempts.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Account is locked due to multiple failed attempts"
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error"
                  error:
                    type: string
                    example: "Internal Server Error"
                  statusCode:
                    type: integer
                    example: 500

  /users/{user_email}/check-lock-status:
    get:
      summary: Check if a user's account is locked.
      description: Returns lock status of the user account.
      operationId: checkLockStatus
      parameters:
        - in: path
          name: user_email
          required: true
          description: email of the user to check.
          schema:
            type: string
            example: "user@gmail.com"
      responses:
        '200':
          description: Lock status retrieved successfully.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/ACCOUNT_LOCKING_MESSAGES"
              examples:
                ACCOUNT_NOT_LOCKED:
                  value:
                    message: Account is not locked
                    isLocked: 0
                    lockedDate: null
                ACCOUNT_LOCKED:
                  value:
                    message: Account is locked
                    isLocked: 1
                    lockedDate: 2025-04-17T07:00:20.549Z
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
                  isLocked:
                    type: integer
                    example: -1
        '500':
          description: Error checking lock status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Error checking lock status"
                  isLocked:
                    type: integer
                    example: -1

components:
  schemas:
    ACCOUNT_LOCKING_MESSAGES:
      type: object
      properties:
        message:
          type: string
        isLocked:
          type: integer
        lockedDate:
            type: string
            format: date-time
            nullable: true
    ERROR_CREATE_USER: 
      type: object
      properties:
        message:
          type: string                       
        error:
          type: string
        statusCode:
          type: integer
    User:
      type: object
      properties:
        uuid:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          example: user@gmail.com
        name:
          type: string
          example: Username
        urlProfilePhoto:
          type: string
          format: uri
          example: https://firebasestorage.googleapis.com/v0/profile_picture_user.jpg
        provider:
          type: string
          example: google.com
        latitude:
          type: number
          format: float
          nullable: true
          example: 34.6037
        longitude:
          type: number
          format: float
          nullable: true
          example: 58.3816
        failedAttempts:
          type: integer
          example: 3
        accountLocked:
          type: boolean
          example: true
        lockUntil:
          type: string
          format: date-time
          nullable: true
          example: "2025-04-13T12:00:00Z"
        lastFailedAt:
          type: string
          format: date-time
          nullable: true
          example: "2025-04-12T14:00:00Z"